Add policy and auth method for blog posts

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,8 @@ class PostController extends Controller
 {
    public function index(){
       //Used of eager loading:
-      $posts = Post::with(['user', 'likes'])->paginate(4);
+      //orderBy('created_at', 'desc') or latest()
+      $posts = Post::orderBy('created_at', 'desc')->with(['user', 'likes'])->paginate(4);
       return view('posts.index', ['posts' => $posts]); 
    }
 
@@ -22,6 +23,13 @@ class PostController extends Controller
    }
 
    public function destroy(Post $post) {
+
+      // if (!$post->ownedBy(auth()->user())) {
+      //    dd('no');
+      // }
+
+      $this->authorize('delete', $post); 
+
       $post->delete();
 
       return back();
